Reorganise: Move logging, json, secrets packages into util folder
