+++ src/internal/handlers/messages/main_handler.go
@@ -7,7 +7,7 @@ import (
 	tgbotapi "github.com/go-telegram-bot-api/telegram-bot-api/v5"
 	"github.com/jonleeyz/bball8bot/commands"
 	customerrors "github.com/jonleeyz/bball8bot/internal/custom-errors"
-	"github.com/jonleeyz/bball8bot/internal/logging"
+	"github.com/jonleeyz/bball8bot/internal/util/logging"
 )
 
 type MessageHandler struct {
